+++ samples/samples-linked/summary_memory_release.cpp
@@ -7,7 +7,7 @@
  * @author mkhan3189
  */
 
-#include "../easylogging++.h"
+#include "../../easylogging++.h"
 
 _INITIALIZE_EASYLOGGINGPP