+++ samples/samples-linked/time_waster.cpp
@@ -6,7 +6,7 @@
  * @author mkhan3189
  */
 
-#include "../easylogging++.h"
+#include "../../easylogging++.h"
 #include <unistd.h>
 
 _INITIALIZE_EASYLOGGINGPP